Create README.md
The online examination system is a web-based application where the educational institutes can conduct the exams in the
online mode where as in the traditional method the students have to go to their colleges, schools or the examination hall so,
the traditional methods conducting examinations takes a long time and it is very difficult for the students to travel when the
examination hall is located too far from their house.
In this we are trying to take online examinations for students. Also registration page for students and add a few pages for
administrators where admin can add questions, see the result of students and see the list of students, also trying to add
subject wise questions.in student module, they have to give the exam generated by admin and see the result at the end of
the test.